feat(dashboard): moved submit logic to form
